CODENVY-198: cleanup modules structure
