+++ assembly/onpremises-ide-packaging-war-platform-api/pom.xml
@@ -123,10 +123,6 @@
             <artifactId>codenvy-router-machine</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.codenvy.plugin</groupId>
-            <artifactId>codenvy-swarm-machine</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
         </dependency>
