+++ plugins/plugin-hosted/codenvy-machine-hosted/pom.xml
@@ -26,10 +26,6 @@
     <name>Codenvy Plugin :: Hosted :: Machine</name>
     <dependencies>
         <dependency>
-            <groupId>com.codenvy.plugin</groupId>
-            <artifactId>codenvy-swarm-machine</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
         </dependency>
@@ -56,6 +52,10 @@
         <dependency>
             <groupId>org.eclipse.che.core</groupId>
             <artifactId>che-core-api-model</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.che.core</groupId>
+            <artifactId>che-core-commons-json</artifactId>
         </dependency>
         <dependency>
             <groupId>org.eclipse.che.core</groupId>
