+++ plugins/plugin-hosted/codenvy-machine-hosted/src/main/java/com/codenvy/swarm/client/RandomNodeSelectionStrategy.java
@@ -12,29 +12,28 @@
  * is strictly forbidden unless prior written permission is obtained
  * from Codenvy S.A..
  */
-package com.codenvy.swarm.client.json;
+package com.codenvy.swarm.client;
+
+import com.codenvy.swarm.client.model.DockerNode;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.Random;
 
 /**
- * Represents node where docker runs.
- * Used for workarounds because of not implemented APIs in Swarm
+ * Random selection strategy for Swarm connector.
  *
  * @author Eugene Voevodin
  */
-public class DockerNode {
-    //TODO add ram and containers ?
-    private final String hostname;
-    private final String addr;
+public class RandomNodeSelectionStrategy implements NodeSelectionStrategy {
 
-    public DockerNode(String hostname, String addr) {
-        this.hostname = hostname;
-        this.addr = addr;
-    }
+    private static final Random random = new Random();
 
-    public String getAddr() {
-        return addr;
-    }
-
-    public String getHostname() {
-        return hostname;
+    @Override
+    public DockerNode select(List<DockerNode> uris) throws IOException {
+        if (uris.size() == 0) {
+            throw new IOException("No docker nodes available");
+        }
+        return uris.get(random.nextInt(uris.size()));
     }
 }
